refactored: baum-welch integrated into MLHMM and deleted
