+++ bhmm/init/__init__.py
@@ -1,1 +1,3 @@
 __author__ = 'noe'
+
+from api import generate_initial_model
